+++ odoo/local-src/sf_rma_mrp_repair/models/__init__.py
@@ -1,1 +1,3 @@
 from . import mrp_repair
+from . import rma
+from . import picking