+++ HISTORY.rst
@@ -28,6 +28,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * BSSFL-352: Add security group to reset RMAs
+* BSSFL-387: Repair state draft and open
 
 **Bugfixes**